+++ src/num/fin/bit/ops.rs
@@ -31,8 +31,8 @@ pub trait BitOps where Self: Sized {
     fn bit_mask_range(start: u32, end: u32) -> Self;
 
     #[doc = _DOC_BIT_MASK_RANGE_CHECKED!()]
-    #[doc = include_str!("_benches/mask_checked_range.md")]
-    fn bit_mask_checked_range(start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = include_str!("_benches/mask_range_checked.md")]
+    fn bit_mask_range_checked(start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_IS_SET_MASK!()]
@@ -53,16 +53,16 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_GET_RANGE!()]
     fn bit_get_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_GET_CHECKED_RANGE!()]
-    fn bit_get_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_GET_RANGE_CHECKED!()]
+    fn bit_get_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     /* get value */
 
     #[must_use]
     #[doc = _DOC_BIT_GET_VALUE_RANGE!()]
     fn bit_get_value_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_GET_VALUE_CHECKED_RANGE!()]
-    fn bit_get_value_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_GET_VALUE_RANGE_CHECKED!()]
+    fn bit_get_value_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     /* set */
 
@@ -81,14 +81,14 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_SET_RANGE!()]
     fn bit_is_set_range(self, start: u32, end: u32) -> bool;
-    #[doc = _DOC_BIT_SET_CHECKED_RANGE!()]
-    fn bit_is_set_checked_range(self, start: u32, end: u32) -> Result<bool, MismatchedBounds>;
+    #[doc = _DOC_BIT_SET_RANGE_CHECKED!()]
+    fn bit_is_set_range_checked(self, start: u32, end: u32) -> Result<bool, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_SET_RANGE!()]
     fn bit_set_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_SET_CHECKED_RANGE!()]
-    fn bit_set_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_SET_RANGE_CHECKED!()]
+    fn bit_set_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     #[doc = _DOC_BIT_SET_ALL!()]
     fn bit_set_all(self) -> Self;
@@ -98,11 +98,11 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_SET_VALUE_RANGE!()]
     fn bit_set_value_range(self, value: Self::Inner, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_SET_VALUE_CHECKED_RANGE!()]
-    fn bit_set_value_checked_range(self, value: Self::Inner, start: u32, end: u32)
+    #[doc = _DOC_BIT_SET_VALUE_RANGE_CHECKED!()]
+    fn bit_set_value_range_checked(self, value: Self::Inner, start: u32, end: u32)
         -> Result<Self, MismatchedBounds>;
-    #[doc = _DOC_BIT_SET_CHECKED_VALUE_CHECKED_RANGE!()]
-    fn bit_set_checked_value_checked_range(self, value: Self::Inner, start: u32, end: u32)
+    #[doc = _DOC_BIT_SET_VALUE_RANGE_CHECKED_STRICT!()]
+    fn bit_set_value_range_checked_strict(self, value: Self::Inner, start: u32, end: u32)
         -> Result<Self, MismatchedBounds>;
 
     /* unset */
@@ -122,15 +122,15 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_UNSET_RANGE!()]
     fn bit_is_unset_range(self, start: u32, end: u32) -> bool;
-    #[doc = _DOC_BIT_UNSET_CHECKED_RANGE!()]
-    fn bit_is_unset_checked_range(self, start: u32, end: u32) -> Result<bool, MismatchedBounds>;
+    #[doc = _DOC_BIT_UNSET_RANGE_CHECKED!()]
+    fn bit_is_unset_range_checked(self, start: u32, end: u32) -> Result<bool, MismatchedBounds>;
 
 
     #[must_use]
     #[doc = _DOC_BIT_UNSET_RANGE!()]
     fn bit_unset_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_UNSET_CHECKED_RANGE!()]
-    fn bit_unset_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_UNSET_RANGE_CHECKED!()]
+    fn bit_unset_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     #[doc = _DOC_BIT_UNSET_ALL!()]
     fn bit_unset_all(self) -> Self;
@@ -146,14 +146,14 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_FLIP_RANGE!()]
     fn bit_flip_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_FLIP_CHECKED_RANGE!()]
-    fn bit_flip_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_FLIP_RANGE_CHECKED!()]
+    fn bit_flip_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_FLIP_RANGE_IF!()]
     fn bit_flip_range_if(self, start: u32, end: u32, cond: bool) -> Self;
-    #[doc = _DOC_BIT_FLIP_CHECKED_RANGE_IF!()]
-    fn bit_flip_checked_range_if(self, start: u32, end: u32, cond: bool)
+    #[doc = _DOC_BIT_FLIP_RANGE_CHECKED_IF!()]
+    fn bit_flip_range_if_checked(self, start: u32, end: u32, cond: bool)
         -> Result<Self, MismatchedBounds>;
 
     /* reverse */
@@ -161,37 +161,37 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_REVERSE_RANGE!()]
     fn bit_reverse_range(self, start: u32, end: u32) -> Self;
-    #[doc = _DOC_BIT_REVERSE_CHECKED_RANGE!()]
-    fn bit_reverse_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
+    #[doc = _DOC_BIT_REVERSE_RANGE_CHECKED!()]
+    fn bit_reverse_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds>;
 
     /* count */
 
     #[must_use]
     #[doc = _DOC_BIT_COUNT_ONES_RANGE!()]
     fn bit_count_ones_range(self, start: u32, end: u32) -> u32;
-    #[doc = _DOC_BIT_COUNT_ONES_CHECKED_RANGE!()]
-    fn bit_count_ones_checked_range(self, start: u32, end: u32) -> Result<u32, MismatchedBounds>;
+    #[doc = _DOC_BIT_COUNT_ONES_RANGE_CHECKED!()]
+    fn bit_count_ones_range_checked(self, start: u32, end: u32) -> Result<u32, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_COUNT_ZEROS_RANGE!()]
     fn bit_count_zeros_range(self, start: u32, end: u32) -> u32;
-    #[doc = _DOC_BIT_COUNT_ZEROS_CHECKED_RANGE!()]
-    fn bit_count_zeros_checked_range(self, start: u32, end: u32) -> Result<u32, MismatchedBounds>;
+    #[doc = _DOC_BIT_COUNT_ZEROS_RANGE_CHECKED!()]
+    fn bit_count_zeros_range_checked(self, start: u32, end: u32) -> Result<u32, MismatchedBounds>;
 
     /* find first */
 
     #[must_use]
     #[doc = _DOC_BIT_FIND_FIRST_ONE_RANGE!()]
     fn bit_find_first_one_range(self, start: u32, end: u32) -> Option<u32>;
-    #[doc = _DOC_BIT_FIND_FIRST_ONE_CHECKED_RANGE!()]
-    fn bit_find_first_one_checked_range(self, start: u32, end: u32)
+    #[doc = _DOC_BIT_FIND_FIRST_ONE_RANGE_CHECKED!()]
+    fn bit_find_first_one_range_checked(self, start: u32, end: u32)
         -> Result<Option<u32>, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_FIND_FIRST_ZERO_RANGE!()]
     fn bit_find_first_zero_range(self, start: u32, end: u32) -> Option<u32>;
-    #[doc = _DOC_BIT_FIND_FIRST_ZERO_CHECKED_RANGE!()]
-    fn bit_find_first_zero_checked_range(self, start: u32, end: u32)
+    #[doc = _DOC_BIT_FIND_FIRST_ZERO_RANGE_CHECKED!()]
+    fn bit_find_first_zero_range_checked(self, start: u32, end: u32)
         -> Result<Option<u32>, MismatchedBounds>;
 
     /* find last */
@@ -199,15 +199,15 @@ pub trait BitOps where Self: Sized {
     #[must_use]
     #[doc = _DOC_BIT_FIND_LAST_ONE_RANGE!()]
     fn bit_find_last_one_range(self, start: u32, end: u32) -> Option<u32>;
-    #[doc = _DOC_BIT_FIND_LAST_ONE_CHECKED_RANGE!()]
-    fn bit_find_last_one_checked_range(self, start: u32, end: u32)
+    #[doc = _DOC_BIT_FIND_LAST_ONE_RANGE_CHECKED!()]
+    fn bit_find_last_one_range_checked(self, start: u32, end: u32)
         -> Result<Option<u32>, MismatchedBounds>;
 
     #[must_use]
     #[doc = _DOC_BIT_FIND_LAST_ZERO_RANGE!()]
     fn bit_find_last_zero_range(self, start: u32, end: u32) -> Option<u32>;
-    #[doc = _DOC_BIT_FIND_LAST_ZERO_CHECKED_RANGE!()]
-    fn bit_find_last_zero_checked_range(self, start: u32, end: u32)
+    #[doc = _DOC_BIT_FIND_LAST_ZERO_RANGE_CHECKED!()]
+    fn bit_find_last_zero_range_checked(self, start: u32, end: u32)
         -> Result<Option<u32>, MismatchedBounds>;
 }
 
@@ -229,8 +229,8 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_mask_range(start: u32, end: u32) -> Self {
                 Bitwise::<$t>::mask_range(start, end).0
             }
-            fn bit_mask_checked_range(start: u32, end: u32) -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise::<$t>::mask_checked_range(start, end)?.0)
+            fn bit_mask_range_checked(start: u32, end: u32) -> Result<Self, MismatchedBounds> {
+                Ok(Bitwise::<$t>::mask_range_checked(start, end)?.0)
             }
 
             #[doc = _DOC_BIT_IS_SET_MASK!()]
@@ -250,8 +250,8 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_get_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).get_range(start, end).0
             }
-            fn bit_get_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).get_checked_range(start, end)?.0)
+            fn bit_get_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds> {
+                Ok(Bitwise(self).get_range_checked(start, end)?.0)
             }
 
             /* get value */
@@ -259,9 +259,9 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_get_value_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).get_value_range(start, end).0
             }
-            fn bit_get_value_checked_range(self, start: u32, end: u32)
+            fn bit_get_value_range_checked(self, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).get_value_checked_range(start, end)?.0)
+                Ok(Bitwise(self).get_value_range_checked(start, end)?.0)
             }
 
             /* set */
@@ -278,16 +278,16 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_is_set_range(self, start: u32, end: u32) -> bool {
                 Bitwise(self).is_set_range(start, end)
             }
-            fn bit_is_set_checked_range(self, start: u32, end: u32)
+            fn bit_is_set_range_checked(self, start: u32, end: u32)
                 -> Result<bool, MismatchedBounds> {
-                Bitwise(self).is_set_checked_range(start, end)
+                Bitwise(self).is_set_range_checked(start, end)
             }
 
             fn bit_set_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).set_range(start, end).0
             }
-            fn bit_set_checked_range(self, start: u32, end: u32) -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).set_checked_range(start, end)?.0)
+            fn bit_set_range_checked(self, start: u32, end: u32) -> Result<Self, MismatchedBounds> {
+                Ok(Bitwise(self).set_range_checked(start, end)?.0)
             }
 
             fn bit_set_all(self) -> Self { Bitwise(self).set_all().0 }
@@ -297,13 +297,13 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_set_value_range(self, value: Self::Inner, start: u32, end: u32) -> Self {
                 Bitwise(self).set_value_range(value, start, end).0
             }
-            fn bit_set_value_checked_range(self, value: Self::Inner, start: u32, end: u32)
+            fn bit_set_value_range_checked(self, value: Self::Inner, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).set_value_checked_range(value, start, end)?.0)
+                Ok(Bitwise(self).set_value_range_checked(value, start, end)?.0)
             }
-            fn bit_set_checked_value_checked_range(self, value: Self::Inner, start: u32, end: u32)
+            fn bit_set_value_range_checked_strict(self, value: Self::Inner, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).set_checked_value_checked_range(value, start, end)?.0)
+                Ok(Bitwise(self).set_value_range_checked_strict(value, start, end)?.0)
             }
 
             /* unset */
@@ -320,17 +320,17 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_is_unset_range(self, start: u32, end: u32) -> bool {
                 Bitwise(self).is_unset_range(start, end)
             }
-            fn bit_is_unset_checked_range(self, start: u32, end: u32)
+            fn bit_is_unset_range_checked(self, start: u32, end: u32)
                 -> Result<bool, MismatchedBounds> {
-                Bitwise(self).is_unset_checked_range(start, end)
+                Bitwise(self).is_unset_range_checked(start, end)
             }
 
             fn bit_unset_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).unset_range(start, end).0
             }
-            fn bit_unset_checked_range(self, start: u32, end: u32)
+            fn bit_unset_range_checked(self, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).unset_checked_range(start, end)?.0)
+                Ok(Bitwise(self).unset_range_checked(start, end)?.0)
             }
 
             fn bit_unset_all(self) -> Self { Bitwise(self).unset_all().0 }
@@ -345,17 +345,17 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_flip_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).flip_range(start, end).0
             }
-            fn bit_flip_checked_range(self, start: u32, end: u32)
+            fn bit_flip_range_checked(self, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).flip_checked_range(start, end)?.0)
+                Ok(Bitwise(self).flip_range_checked(start, end)?.0)
             }
 
             fn bit_flip_range_if(self, start: u32, end: u32, cond: bool) -> Self {
                 Bitwise(self).flip_range_if(start, end, cond).0
             }
-            fn bit_flip_checked_range_if(self, start: u32, end: u32, cond: bool)
+            fn bit_flip_range_if_checked(self, start: u32, end: u32, cond: bool)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).flip_checked_range_if(start, end, cond)?.0)
+                Ok(Bitwise(self).flip_range_if_checked(start, end, cond)?.0)
             }
 
             /* reverse */
@@ -363,9 +363,9 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_reverse_range(self, start: u32, end: u32) -> Self {
                 Bitwise(self).reverse_range(start, end).0
             }
-            fn bit_reverse_checked_range(self, start: u32, end: u32)
+            fn bit_reverse_range_checked(self, start: u32, end: u32)
                 -> Result<Self, MismatchedBounds> {
-                Ok(Bitwise(self).reverse_checked_range(start, end)?.0)
+                Ok(Bitwise(self).reverse_range_checked(start, end)?.0)
             }
 
             /* count */
@@ -373,16 +373,16 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_count_ones_range(self, start: u32, end: u32) -> u32 {
                 Bitwise(self).count_ones_range(start, end)
             }
-            fn bit_count_ones_checked_range(self, start: u32, end: u32)
+            fn bit_count_ones_range_checked(self, start: u32, end: u32)
                 -> Result<u32, MismatchedBounds> {
-                Bitwise(self).count_ones_checked_range(start, end)
+                Bitwise(self).count_ones_range_checked(start, end)
             }
             fn bit_count_zeros_range(self, start: u32, end: u32) -> u32 {
                 Bitwise(self).count_zeros_range(start, end)
             }
-            fn bit_count_zeros_checked_range(self, start: u32, end: u32)
+            fn bit_count_zeros_range_checked(self, start: u32, end: u32)
                 -> Result<u32, MismatchedBounds> {
-                Bitwise(self).count_zeros_checked_range(start, end)
+                Bitwise(self).count_zeros_range_checked(start, end)
             }
 
             /* find first */
@@ -390,16 +390,16 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_find_first_one_range(self, start: u32, end: u32) -> Option<u32> {
                 Bitwise(self).find_first_one_range(start, end)
             }
-            fn bit_find_first_one_checked_range(self, start: u32, end: u32)
+            fn bit_find_first_one_range_checked(self, start: u32, end: u32)
                 -> Result<Option<u32>, MismatchedBounds> {
-                Bitwise(self).find_first_one_checked_range(start, end)
+                Bitwise(self).find_first_one_range_checked(start, end)
             }
             fn bit_find_first_zero_range(self, start: u32, end: u32) -> Option<u32> {
                 Bitwise(self).find_first_zero_range(start, end)
             }
-            fn bit_find_first_zero_checked_range(self, start: u32, end: u32)
+            fn bit_find_first_zero_range_checked(self, start: u32, end: u32)
                 -> Result<Option<u32>, MismatchedBounds> {
-                Bitwise(self).find_first_zero_checked_range(start, end)
+                Bitwise(self).find_first_zero_range_checked(start, end)
             }
 
             /* find last */
@@ -407,16 +407,16 @@ macro_rules! _num_fin_bit_ops_impl_fns {
             fn bit_find_last_one_range(self, start: u32, end: u32) -> Option<u32> {
                 Bitwise(self).find_last_one_range(start, end)
             }
-            fn bit_find_last_one_checked_range(self, start: u32, end: u32)
+            fn bit_find_last_one_range_checked(self, start: u32, end: u32)
                 -> Result<Option<u32>, MismatchedBounds> {
-                Bitwise(self).find_last_one_checked_range(start, end)
+                Bitwise(self).find_last_one_range_checked(start, end)
             }
             fn bit_find_last_zero_range(self, start: u32, end: u32) -> Option<u32> {
                 Bitwise(self).find_last_zero_range(start, end)
             }
-            fn bit_find_last_zero_checked_range(self, start: u32, end: u32)
+            fn bit_find_last_zero_range_checked(self, start: u32, end: u32)
                 -> Result<Option<u32>, MismatchedBounds> {
-                Bitwise(self).find_last_zero_checked_range(start, end)
+                Bitwise(self).find_last_zero_range_checked(start, end)
             }
         }
     };