+++ docs/changelog.md
@@ -184,7 +184,9 @@
 - make `Float`'s std methods const: `ceil`, `floor`, `mul_add`, `round_ties_away`, `round_ties_even`.
 
 #### num::fin::bit
-- fix `Bitwise::set_checked_value_checked_range`.
+- update `Bitwise`:
+  - rename checked methods to have `_checked` suffix.
+  - rename `set_checked_value_checked_range` to `set_value_range_checked_strict` and fix.
 
 ### num::grain
 - rename `Primitive*` traits to `Prim*`.